fix:fix error in product float to float64

+++ product-service/backend-api/models/product.go
@@ -5,5 +5,5 @@ type Product struct {
 	Title   string `json:"title"`
 	Content string `json:"content"`
 	Image   string `json:"image"`
-	Price	float  `json:"Price"`
+	Price	float64  `json:"Price"`
 }
